util: add ShouldBeRetimed annotation
